#1821 Separate optional modules by jdk versions Splits into 3 modules - jdk6, 7, and 8 so that each module is compiled against their specific jdk versions. Also moves test codes in the java8-test module into their respective optional packages.
