--- java8-test/.gitignore
@@ -1,5 +1,0 @@
-/.settings/
-/*.iml
-/target/
-/.classpath
-/.project
