--- java8-test/clover.license
@@ -1,5 +1,0 @@
-RMRqrdbgbKFhbaVnDxHUdDQvrOQXxIBklnvcmahheubVC
-mh2KM35CLkwUHS4DH7QVhxy52J5hnWbyEm6Cyd3KkF<mV
-RmmnSVOqOMnOnMMrmMqwXomoroNrqPNRrPSsWwtUxXuUU
-sRONqpnmqmUUnqonmstsmmmmmUUnqonmstsmmmmmUUGfk
-mlfkqUUnmmmm
