--- java8-test/src/test/java/com/navercorp/test/pinpoint/jdk8/lambda/Mozzi.java
@@ -1,32 +1,0 @@
-/*
- * Copyright 2014 NAVER Corp.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.navercorp.test.pinpoint.jdk8.lambda;
-
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class Mozzi {
-    private final int age;
-    
-    public Mozzi(int age) {
-        this.age = age;
-    }
-
-    public int getAge() {
-        return age;
-    }
-}
