+++ profiler-optional/pom.xml
@@ -8,36 +8,21 @@
     
     <artifactId>pinpoint-profiler-optional</artifactId>
     <name>pinpoint-profiler-optional</name>
+    <packaging>pom</packaging>
     <description>pinpoint profiler optional package</description>
-    
-    <properties>
-        <jdk.version>1.7</jdk.version>
-        <jdk.home>${env.JAVA_7_HOME}</jdk.home>
-    </properties>
-    
+
+    <modules>
+        <module>profiler-optional-jdk6</module>
+        <module>profiler-optional-jdk7</module>
+        <module>profiler-optional-jdk8</module>
+        <module>profiler-optional-parent</module>
+    </modules>
+
     <dependencies>
         <dependency>
             <groupId>com.navercorp.pinpoint</groupId>
-            <artifactId>pinpoint-profiler</artifactId>
-            <scope>provided</scope>
-            <exclusions>
-                <exclusion>
-                    <groupId>com.navercorp.pinpoint</groupId>
-                    <artifactId>pinpoint-commons</artifactId>
-                </exclusion>
-                <exclusion>
-                    <groupId>com.navercorp.pinpoint</groupId>
-                    <artifactId>pinpoint-bootstrap</artifactId>
-                </exclusion>
-                <exclusion>
-                    <groupId>com.navercorp.pinpoint</groupId>
-                    <artifactId>pinpoint-rpc</artifactId>
-                </exclusion>
-                <exclusion>
-                    <groupId>com.navercorp.pinpoint</groupId>
-                    <artifactId>pinpoint-thrift</artifactId>
-                </exclusion>
-            </exclusions>
+            <artifactId>pinpoint-profiler-optional-jdk7</artifactId>
+            <version>${project.version}</version>
         </dependency>
     </dependencies>
 </project>
