+++ profiler-optional/profiler-optional-jdk6/.gitignore
@@ -1,0 +1,5 @@
+/.settings/
+/*.iml
+/.project
+/target/
+/.classpath
