+++ profiler-optional/profiler-optional-jdk6/pom.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>com.navercorp.pinpoint</groupId>
+        <artifactId>pinpoint-profiler-optional-parent</artifactId>
+        <relativePath>../profiler-optional-parent</relativePath>
+        <version>1.6.0-SNAPSHOT</version>
+    </parent>
+
+    <artifactId>pinpoint-profiler-optional-jdk6</artifactId>
+    <name>pinpoint-profiler-optional-jdk6</name>
+    <packaging>jar</packaging>
+    <description>pinpoint profiler optional package for jdk 6</description>
+
+    <properties>
+        <jdk.version>1.6</jdk.version>
+        <jdk.home>${env.JAVA_6_HOME}</jdk.home>
+    </properties>
+
+</project>
