+++ profiler-optional/profiler-optional-jdk7/.gitignore
@@ -1,0 +1,5 @@
+/.settings/
+/*.iml
+/.project
+/target/
+/.classpath
