+++ profiler-optional/profiler-optional-jdk7/src/main/java/com/navercorp/pinpoint/profiler/monitor/codahale/cpu/metric/EnhancedCpuLoadMetricSet.java
@@ -1,0 +1,52 @@
+/*
+ * Copyright 2014 NAVER Corp.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.navercorp.pinpoint.profiler.monitor.codahale.cpu.metric;
+
+import com.codahale.metrics.Gauge;
+import com.sun.management.OperatingSystemMXBean;
+
+/**
+ * @author hyungil.jeong
+ */
+public class EnhancedCpuLoadMetricSet extends AbstractCpuLoadMetricSet {
+
+    @Override
+    protected Gauge<Double> getJvmCpuLoadGauge(final OperatingSystemMXBean operatingSystemMXBean) {
+        return new Gauge<Double>() {
+            @Override
+            public Double getValue() {
+                return operatingSystemMXBean.getProcessCpuLoad();
+            }
+        };
+    }
+
+    @Override
+    protected Gauge<Double> getSystemCpuLoadGauge(final OperatingSystemMXBean operatingSystemMXBean) {
+        return new Gauge<Double>() {
+            @Override
+            public Double getValue() {
+                return operatingSystemMXBean.getSystemCpuLoad();
+            }
+        };
+    }
+
+    @Override
+    public String toString() {
+        return "CpuLoadMetricSet for Java 1.7+";
+    }
+
+}
