+++ profiler-optional/profiler-optional-jdk7/src/test/java/bug_regression_jdk7/javassist/CustomURLClassLoader.java
