+++ profiler-optional/profiler-optional-jdk7/src/test/resources/log4j.xml
