+++ profiler-optional/profiler-optional-jdk8/.gitignore
@@ -1,0 +1,5 @@
+/.settings/
+/*.iml
+/.project
+/target/
+/.classpath
