+++ profiler-optional/profiler-optional-jdk8/src/test/resources/lambda-test.xml
