+++ profiler-optional/profiler-optional-jdk8/src/test/resources/pinpoint-lambda-test.config
